+++ ci/travis-sync-master-canary.sh
@@ -1,5 +1,8 @@
 echo "chore: syncing master to canary"
 git fetch origin canary:canary
+echo "Checking out to canary"
 git checkout canary
+echo "Merging Master to Canary"
 git merge master -m 'chore: syncing master to canary [skip ci]'
+echo "Pushing Sync"
 git push origin canary